Move GET /api/v1/accounts/:id specs to accounts_show_spec.rb
Wrap the existing specs in an "endpoint behavior" describe to leave
room for a parallel "response keys" describe in subsequent commits.
